requirements.txt
requests
python-dotenv
flask
waitress
flask-cors
